// HtmlEditor/HtmlEditorComponent.d.ts
import * as React from "react";
export interface HtmlEditorProps {
    htmlContent: string;
    onHtmlChange: (htmlContent: string) => void;
    disabled: boolean;
}
export declare const HtmlEditorComponent: (props: HtmlEditorProps) => React.JSX.Element;
//# sourceMappingURL=HtmlEditorComponent.d.ts.map